requirements.txt
# Core dependencies for RAG framework and Llama Index
llama-index-core==0.11.0
llama-index-embeddings-gemini
llama-index-embeddings-huggingface
llama-index-llms-gemini
llama-index-llms-openai
llama-index-llms-ollama
llama-index-readers-file

# Vector store integrations
faiss-cpu
llama-index-vector-stores-faiss  

# Retrieval and ranking
rank-bm25

# Data collection and parsing
arxiv
requests
beautifulsoup4

# Document Processing
pypdf

# Additional utilities
python-dotenv

# UI dependencies - Gradio 6.x
gradio==6.0.2
gradio_client==2.0.1
